feat: add declare-v2
