+++ Cargo.toml
@@ -7,7 +7,6 @@ include = ["/base"]
 [lib]
 name = "nile_rs"
 path = "src/nile-rs/lib.rs"
-doc = false
 
 [[bin]]
 name = "nile-rs"
@@ -17,19 +16,18 @@ doc = false
 anyhow = "1.0.69"
 async-trait = "0.1.64"
 clap = { version = "4.1.4", features = ["derive"] }
+cairo-lang-starknet = "1.0.0-alpha.6"
 dotenv = "0.15.0"
 figment = { version = "0.10", features = ["toml", "env", "test"] }
-cairo-lang-compiler = "1.0.0-alpha.3"
-cairo-lang-starknet = "1.0.0-alpha.3"
-scarb = "0.1.0-rc.2"
+scarb = "0.1.0"
 serde = "1.0.152"
 serde_json = "1.0.93"
-starknet-accounts = { git = "https://github.com/xJonathanLEI/starknet-rs"}
-starknet-core = { git = "https://github.com/xJonathanLEI/starknet-rs"}
-starknet-contract = { git = "https://github.com/xJonathanLEI/starknet-rs"}
-starknet-crypto = { git = "https://github.com/xJonathanLEI/starknet-rs"}
-starknet-providers = { git = "https://github.com/xJonathanLEI/starknet-rs"}
-starknet-signers = { git = "https://github.com/xJonathanLEI/starknet-rs"}
+starknet-accounts = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet-core = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet-contract = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet-crypto = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet-providers = { git = "https://github.com/xJonathanLEI/starknet-rs" }
+starknet-signers = { git = "https://github.com/xJonathanLEI/starknet-rs" }
 tokio = { version = "1", features = ["full"] }
 url = "2.3.1"
 walkdir = "2.3.2"
